requirements.txt
Django>=6.0,<6.1
Pillow>=10.0
qrcode>=8.0
nanoid>=2.0
django-simple-history>=3.7
whitenoise>=6.8
celery>=5.4
django-auth-ldap>=5.1
python-ldap>=3.4
gunicorn>=23.0
python-dotenv>=1.0
